Test sample TA with DES - Fails badly!
I had some issues using OPTEE with DES encryption. AES worked for me.
To be sure that the problem is not part of the software I'm working on,
I pulled the sample TA and created this patch to make it try to use
DES - with the same effect: Calling TEE_CiperInitialize() leads to a
Panic(0xffff0007) - "TEEC_ERROR_BAD_STATE".

To reproduce, compile the TA and CA and run

    ~# sampleca cbc enc clear crypt

When trying to use a 64 bit key with DES, I get a regular error - which
is expected behaviour.
What and if I specify as IV with TEE_CipherInit seems to be totally
unimportant.

Signed-off-by: Philipp Hachtmann <phachtmann@de.adit-jv.com>